Here's the error message I get
When using the combatSystem(hero, tutorialEnemy) function I get this error message
What is supposed to happen: 
The program determines the loser by finding the first party with 0 health. The loser is then put into a variable called loser. 
If the player's name is inside the loser variable, then it brings them to the restart menu.
If the player's name is not inside the loser variable, then the program decides that the enemy has been slain and the player gains rewards.